infrastructure/requirements.txt
aws-cdk-lib>2.103.0,<3.0
constructs>=10.0.0
aws-cdk.aws_apigatewayv2_integrations_alpha
aws-cdk.aws_apigatewayv2_alpha
cdk-ecr-deployment